cli: move enumerate_configs to sopel.cli.utils
