+++ sopel/cli/__init__.py
@@ -1,0 +1,5 @@
+# coding=utf-8
+from __future__ import unicode_literals, absolute_import, print_function, division
+
+# Shortcut imports
+from .utils import enumerate_configs  # noqa
